Dependencies would be slightly clearer if you were to combine all the import code blocks into one.
==> Done

In create_tmpfile, is there a chance it could be called twice with the same file name? If so, is it OK for shutil.copy() to overwrite the   previous one? I'm assuming that's the case, and if so, no change needed.

==> The caller of this function is responsible to give unique directory names.
==> The code just deletes any existing files in that directory before creating the new one.
==> Yes the overwrite functionality is assumed.

In MammalFind, please set self.root_dir = orca_params.DATA_PATH (instead of hard-coding)

==> Done

In MammalFind.init()__ you need to exactly match the logic for how lable names were created initially. Also, do we really need to load all  file names into memory?

Your logic in MammalFind._extract_path_segments() doesn't exactly match the logic we use in database_parser.py (look for the regex). But    re: my previous point, I'm not sure you actually need this function to crawl the directory structure. It looks like you've also got similar logic in get_sample().

==> Switched to using the same code as in database_parser.py to load in the file associations.
==> I do like to build this association as I can sample from different waveforms in different iterations.

The notebook could use a little more text (for the benefit of the class watching us scroll through it). You should probably also move most  (all?) of the functions to a supporting .py file you import so that the class can focus more on the high-level flow and audio/visual        elements. Then it will feel more like a web page and less like a bunch of code. I don't think we want to spend time explaining code to      people during the presentation.

==> Agreed. Added the file demo_helper.py and moved most of the code there.

You're using n_mels=128. If that refers to number of frequency bands, you should set it euqla to mel_params.NUM_BANDS (which = 64 in the    real model).

==> Corrected as suggested.

In display_wave(file), when you call librosa.load(file), then librosa.display.wavelot(y, sr=16000), does that actually resample the file up or down in order to get it to 16000? Not all files are recorded at 16k, so if "sr=16000" doesn't trigger automated resampling, then you     probably need to add that step.

==> Yes it resamples as per the `sr` argument.

How did you decide on probability_threshold = 0.6? The numbers I've included in our final report are from the sklearn classification report which just uses 0.5. The default we use in the live_stream_inference.py is orca_params.LIVE_FEED_MINIMUM_INFERENCE_PROBABILITY (0.75) to    reduce false positives, so you might want to use that.

==> Done